+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 11.2.2
 -----------------
 - Hide list count in ReportCards.fnc.php, Transcripts.fnc.php & Preferences.php
+- Security: SQL prevent INSERT or UPDATE on any table in MassCreateAssignments.php, SchoolFields.php, AddressFields.php, PeopleFields.php, StudentFields.php & UserFields.php
 
 Changes in 11.2.1
 -----------------